Fix/330 bug disable spoofing (#335)
* fix / comp operator for spoofing disable

* fix / error handling

+++ backend/src/chatBlockUser/chatBlockUser.service.ts
@@ -44,6 +44,7 @@ export class ChatBlockUserService {
         blockedUserId: param.blockedUserId,
       },
     });
+    if (!targetRecord) return;
     this.chatBlockRepository.remove(targetRecord);
   }
 }
