+++ backend/src/users/users.service.ts
@@ -214,6 +214,7 @@ export class UsersService {
       },
     });
 
+    if (!pendingItem) return;
     this.pendingRepository.remove(pendingItem);
   }
